pyproject.toml
[project]
name = "homework"
version = "0.1.0"
description = "Nord Security Homework Task for Software Engineer"
readme = "README.md"
requires-python = ">=3.13.2"
dependencies = [
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
    "pre-commit>=4.1.0",
    "scrapy>=2.12.0",
]
